// src/models/IAnimals.ts
export interface IAnimal {
  id: number;
  name: string;
  yearOfBirth: string;
  shortDescription: string;
  imageUrl: string;
  longDescription: string;
  lastFed: string;
  isFed: boolean;
}

// export interface ExtendedAnimal {
//   id: string;
//   name: string;
//   yearOfBirth: string;
//   longDescription: string;
//   imageUrl: string;
//   lastFed: string;
//   isFed: boolean;
// }